Move is_range util to range.hpp to avoid Boost.Range/MPL dependency in type_traits.hpp
