+++ include/boost/geometry/util/range.hpp
@@ -21,11 +21,13 @@
 #include <boost/concept_check.hpp>
 #include <boost/config.hpp>
 #include <boost/core/addressof.hpp>
+#include <boost/mpl/has_xxx.hpp>
 #include <boost/range/concepts.hpp>
 #include <boost/range/begin.hpp>
 #include <boost/range/end.hpp>
 #include <boost/range/empty.hpp>
 #include <boost/range/difference_type.hpp>
+#include <boost/range/has_range_iterator.hpp>
 #include <boost/range/iterator.hpp>
 #include <boost/range/rbegin.hpp>
 #include <boost/range/reference.hpp>
@@ -35,9 +37,44 @@
 #include <boost/geometry/core/assert.hpp>
 #include <boost/geometry/core/mutable_range.hpp>
 
-namespace boost { namespace geometry { namespace range {
+namespace boost { namespace geometry { namespace range
+{
 
-namespace detail {
+namespace detail
+{
+
+BOOST_MPL_HAS_XXX_TRAIT_DEF(iterator_category)
+
+template <typename T>
+struct is_iterator
+    : std::integral_constant
+        <
+            bool,
+            has_iterator_category
+                <
+                    std::iterator_traits<T>
+                >::value
+        >
+{};
+
+
+template <typename T, bool HasIterator = boost::has_range_iterator<T>::value>
+struct is_range_impl
+    : is_iterator
+        <
+            typename boost::range_iterator<T>::type
+        >
+{};
+template <typename T>
+struct is_range_impl<T, false>
+    : std::false_type
+{};
+
+template <typename T>
+struct is_range
+    : is_range_impl<T>
+{};
+
 
 // NOTE: For SinglePassRanges pos could iterate over all elements until the i-th element was met.
